backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.1.0
python-docx==1.1.2
